+++ packages/adminui-pub/adminui.js
@@ -3,7 +3,6 @@
  */
 /* global SystemJS */
 const UB = require('@unitybase/ub-pub')
-const Q = require('bluebird-q')
 const _ = require('lodash')
 const csShared = require('@unitybase/cs-shared')
 const fileSaver = require('file-saver')
@@ -12,7 +11,6 @@ const fileSaver = require('file-saver')
 // System.import('lodash') somewhere in code exists
 // Important to use SystemJS instead of System to prevent webpack to optimize calls to System
 if (!SystemJS.has('lodash')) SystemJS.set('lodash', SystemJS.newModule(_))
-if (!SystemJS.has('bluebird-q')) SystemJS.set('bluebird-q', SystemJS.newModule(Q)) // Q Promises hack
 if (!SystemJS.has('@unitybase/cs-shared')) SystemJS.set('@unitybase/cs-shared', SystemJS.newModule(csShared))
 if (!SystemJS.has('file-saver')) SystemJS.set('file-saver', SystemJS.newModule(fileSaver))
 
@@ -25,32 +23,7 @@ window.UB = UB
 window._ = _
 window.UB.LocalDataStore = LocalDataStore
 
-window.Q = Q
 window.UBDomain = UBDomain // used as UBDomain.getPhysicalDataType && UBDomain.ubDataTypes
-
-if (!Promise.prototype.fin) { // Q Promises hack
-  // eslint-disable-next-line no-extend-native
-  Promise.prototype.fin = function (cb) {
-    const res = () => this
-    const fin = () => Promise.resolve(cb()).then(res)
-    return this.then(fin, fin)
-  }
-}
-
-if (!Promise.prototype.finally) { // winXP hack
-  // eslint-disable-next-line no-extend-native
-  Promise.prototype.finally = function (cb) {
-    const res = () => this
-    const fin = () => Promise.resolve(cb()).then(res)
-    return this.then(fin, fin)
-  }
-}
-
-if (!Promise.prototype.done) { // Q Promises hack
-  // eslint-disable-next-line no-extend-native
-  Promise.prototype.done = Promise.prototype.then
-}
-
 // commented by MPV on 2019-03-01 - required in ubs
 // window.JSZip = require('jszip/dist/jszip.js') // for xlsx-pub. require('jszip') dose not work
 window.saveAs = fileSaver.saveAs